pyproject.toml
[project]
name = "my-iptv"
version = "1.0.0"
description = "Malaysian IPTV web application with FastAPI backend"
authors = [
    {name = "Malaysian IPTV", email = "info@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.9.1",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "requests>=2.31.0",
    "lxml>=4.9.3",
    "python-dateutil>=2.8.2",
]

[project.urls]
Homepage = "https://github.com/yourusername/my-iptv"
Repository = "https://github.com/yourusername/my-iptv"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
